Always create a new session when creating a new window
I am not sure of any case where you would want to create a new window and not start a new session? This is changed behavior and makes some plugins that created a new window appear to be broken (a window opens, but you can't interact until you create a new tab).