Support XDG and other user config dirs
Tries to find a config file in the following order:

- $XDG_CONFIG_HOME, if it's set
- Electron's app.getPath('userData'). This is:
  - %APPDATA% on Windows
  - $XDG_CONFIG_HOME on Linux
  - ~/Library/Application Support on macOS
- User's home directory

For all but the home directory, files are in a subfolder of the
directory which is named after the packageName in app/package.json.

For creation of a new config, again, $XDG_CONFIG_HOME is preferred, if
set. Otherwise, it will use the app.getPath('userData') path.