Add category option under linux key in root package.json
In response to issue entry #2097, a specific category option has been defined under the linux key in the root package.json file. This allows Hyper to take advantage of features that have been added to the most recent version of electron-builder.

Note: I tried to make this commit before, but mistakenly added app/yarn.lock, in addition to the solution of the stated issue, to the branch. Thus, this is a corrected version of my earlier attempt.