Make CLI perform static analysis of config lazily
zeit/hyper#2375 adds a wrapper CLI tool that parses the ~/.hyper.js
configuration file. However, the static analysis logic can't possibly
handle every possible configuration file without executing it.

In my case, [my configuration file][0] defines a constant at the top,
which means that the `module.exports` definition isn't the first
statement. This breaks the parsing logic.

However, this static analysis is only necessary if one wishes to use the
CLI subcommands. They're not needed if one just wants to launch hyper.

Unfortunately, the existing implementation tries to read from the AST
immediately when `cli/api.js` is imported. This changes it to be lazy,
and uses memoization retain the same reference, allowing the AST to
still be mutated in-place.

[0]: https://github.com/bgw/dotfiles/blob/665e83fd7a13be675ad0ae984478d5b0de502ee1/hyper.js#L4