perf: improve performance for mapping keys and generating prefixed comamnd
The `generatePrefixedCommand` is a simple pure function. But it used `map()` for looping over the shortcuts. The `map()` creates a new every time and creates unnecessary memory allocations. I have switched to using the `for..loop`.

Another change I made is in the default function for getting keymaps. I have switched from using the method `Object.keys()` which creates a new array and instead used `for..in` loop. I also declared the `keymap` variable outside of the scope to limit the number of times object is modified. Used `for..of` instead of `forEach` because it creates a new function for every iteration.